Fix modal so that dialog
